refactor(dashboard): 将 Memory/Activity 子页合并至 Home/Dashboard 整宽日志面板;
- 动机：Memory/Activity 实质承载平台 Toast 通知历史（localStorage 数据源、跨模块写入），与 Memory 领域错位；本次按正交分解迁回 Dashboard，并将 Memory 二级导航从 7 个 tab 精简为 6 个；
- Dashboard：新建 `_components/ActivityLogPanel.tsx`，复用 ExecutionTimeline 卡片视觉范式（uppercase 头部条 + max-h-[480px] 内滚动 + 5 个 level 过滤 pill / 计数 / Refresh / Clear All）；面板根节点带 `data-testid="activity-log-panel"` 便于 e2e 隔离；
- Hook 上移：`features/memory/hooks/useActivityLog.ts` → `hooks/useActivityLog.ts`，与 useSubAgentsList / useHeartbeatPoll 等平台级 hook 同级；类型 `ActivityEntry`/`ActivityLevel` 自 hook 末尾 re-export；
- 下线：删除 `app/memory/activity/page.tsx` 子页与 MemoryNav 中的 Activity tab；清理 `features/memory/index.ts` 中无领域归属的 useActivityLog / Activity Log Types 区段；
- 测试：删除 `tests/e2e/memory/activity.spec.ts`，新建 `tests/e2e/dashboard/dashboard-activity.spec.ts` 移植 4 个核心断言（空态 / Level 过滤 / Clear All / 损坏 localStorage 降级），并通过 `data-testid` 缩域避免与 ExecutionTimeline 选择器冲突；`memory-pages.spec.ts` 中 7 tab 断言降为 6；
- 文档：`docs/memory/user-guide/basics.md` 同步 UI 导航表格 + 迁移注解；`docs/agents/issue.md` 新增 ISSUE-090 沉淀本次跨模块归属决策；
- 验证：typecheck / lint / 5 个 Memory 相关 unit 测试全绿；浏览器实机验证待用户重启本地 UI 后自查。
