+++ apps/negentropy-ui/app/(home)/dashboard/page.tsx
@@ -2,6 +2,7 @@
 
 import { useCallback, useState } from "react";
 
+import { ActivityLogPanel } from "./_components/ActivityLogPanel";
 import { DimensionCharts } from "./_components/DimensionCharts";
 import { ExecutionTimeline } from "./_components/ExecutionTimeline";
 import { FilterBar } from "./_components/FilterBar";
@@ -32,6 +33,7 @@ const INITIAL_FILTERS: DashboardFilters = {
  * - 30s 兜底定时刷新（防 SSE 抖动）；
  * - 点击任务行 → Drawer 打开 → 加载 /scheduler/tasks/{id}；
  * - Run Now / Toggle 操作完成后调用 refresh() 同步状态。
+ * - 客户端 Activity 面板渲染 localStorage Toast 历史（与后端 Execution Timeline 正交）。
  */
 export default function DashboardPage() {
   const [filters, setFilters] = useState<DashboardFilters>(INITIAL_FILTERS);
@@ -92,6 +94,9 @@ export default function DashboardPage() {
         <div className="lg:col-span-5">
           <ExecutionTimeline executions={executions} />
         </div>
+      </div>
+      <div className="mt-3">
+        <ActivityLogPanel />
       </div>
       <TaskDetailDrawer task={selectedTask} onClose={handleClose} onTaskChanged={refresh} />
     </div>
