+++ apps/negentropy-ui/components/ui/MemoryNav.tsx
@@ -12,7 +12,6 @@ const NAV_ITEMS = [
   { href: "/memory/audit", label: "Audit" },
   { href: "/memory/conflicts", label: "Conflicts" },
   { href: "/memory/automation", label: "Automation" },
-  { href: "/memory/activity", label: "Activity" },
 ];
 
 export function MemoryNav({
