+++ apps/negentropy-ui/features/memory/index.ts
@@ -38,12 +38,6 @@ export type {
   UseMemoryAuditReturnValue,
 } from "./hooks/useMemoryAudit";
 
-export { useActivityLog } from "./hooks/useActivityLog";
-export type {
-  UseActivityLogOptions,
-  UseActivityLogReturnValue,
-} from "./hooks/useActivityLog";
-
 // ============================================================================
 // Components
 // ============================================================================
@@ -81,15 +75,6 @@ export {
   submitRetrievalFeedback,
   fetchRetrievalMetrics,
 } from "./utils/memory-api";
-
-// ============================================================================
-// Activity Log Types
-// ============================================================================
-
-export type {
-  ActivityEntry,
-  ActivityLevel,
-} from "@/lib/activity-store";
 
 // ============================================================================
 // Types
