+++ apps/negentropy-ui/tests/e2e/memory/memory-pages.spec.ts
@@ -274,7 +274,7 @@ test("Conflicts 页面点击冲突项显示详情", async ({ page }) => {
 // Navigation
 // ============================================================================
 
-test("Memory 导航栏包含所有 7 个页面标签", async ({ page }) => {
+test("Memory 导航栏包含所有 6 个页面标签", async ({ page }) => {
   await mockAuthenticatedUser(page);
 
   await page.route("**/api/memory/dashboard**", async (route) => {
@@ -302,5 +302,4 @@ test("Memory 导航栏包含所有 7 个页面标签", async ({ page }) => {
   await expect(page.getByRole("link", { name: "Audit" })).toBeVisible();
   await expect(page.getByRole("link", { name: "Conflicts" })).toBeVisible();
   await expect(page.getByRole("link", { name: "Automation" })).toBeVisible();
-  await expect(page.getByRole("link", { name: "Activity" })).toBeVisible();
 });
